requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bidict==0.23.1
blinker==1.7.0
cantinaUtils==0.0.6
cffi==1.16.0
click==8.1.7
Flask==3.0.2
Flask-SocketIO==5.3.6
h11==0.14.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.21
PyMySQL==1.1.1
pyotp==2.9.0
pypng==0.20220715.0
python-engineio==4.9.0
python-socketio==5.11.1
qrcode==7.4.2
simple-websocket==1.0.0
typing_extensions==4.10.0
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.1
